+++ CHANGELOG.md
@@ -82,6 +82,7 @@ Compatibility:
 * Implement `Thread#native_thread_id` method (#2733, @horakivo).
 * Modify `Struct#{inspect,to_s}` to match MRI when the struct is nested inside of an anonymous class or module (@st0012, @nirvdrum).
 * `Fiber.current` and `Fiber#transfer` are available without `require 'fiber'` like in CRuby 3.1 (#2733, @eregon).
+* Add `freeze` keyword argument to `Marshal.load` (#2733, @andrykonchin).
 
 Performance: